+++ vsu-forum-backend/src/models/post.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Post {
-    pub id: i32,
-    pub topic_id: i32,
-    pub author_id: i32,
+    pub id: i64,
+    pub topic_id: i64,
+    pub author_id: i64,
     pub text: String,
 }
